Introduce decoder base classes.
